+++ spec/dummy/app/controllers/public_controller.rb
@@ -1,2 +1,2 @@
-class PublicController < ApplicationController
+class PublicController < PgEngine::BasePublicController
 end