+++ spec/dummy/app/controllers/users_controller.rb
@@ -1,0 +1,2 @@
+class UsersController < PgEngine::BaseUsersController
+end