+++ spec/dummy/config/routes.rb
@@ -1,7 +1,7 @@
 include PgEngine::RouteHelpers
 
 Rails.application.routes.draw do
-  namespace :frontend, path: 'u' do
+  namespace :users, path: 'u' do
     pg_resource(:cosas)
     pg_resource(:categoria_de_cosas) do
       pg_resource(:cosas)
@@ -13,7 +13,7 @@ Rails.application.routes.draw do
       pg_resource(:cosas)
     end
   end
-  root "frontend/categoria_de_cosas#index"
+  root to: 'users#home'
 
   get :action_with_redirect, to: 'dummy_base#action_with_redirect'
   get :check_dev_user, to: 'dummy_base#check_dev_user'